+++ wire/msgmerkleblock.go
@@ -143,7 +143,7 @@ func (msg *MsgMerkleBlock) Command() string {
 
 // MaxPayloadLength returns the maximum length the payload can be for the
 // receiver.  This is part of the Message interface implementation.
-func (msg *MsgMerkleBlock) MaxPayloadLength(pver uint32) uint32 {
+func (msg *MsgMerkleBlock) MaxPayloadLength(pver uint32) uint64 {
 	return MaxBlockPayload
 }
